REF: replaces HTTP-based log download code with broker-based
This allows removing a lot of very old HTTP code and simplifies the front-end code base. Including:

- report resolvers are not needed w/ the broker downloader
- old server wrapper in the Maestro CLI

It also replaces the weak time-of-expiry based downloader code with use a stale + lock mechanism to control the wait for download
