+++ maestro-reports/src/main/java/org/maestro/reports/downloaders/DownloaderBuilder.java
@@ -31,13 +31,8 @@ public class DownloaderBuilder {
         }
 
         final ReportsDownloader reportsDownloader;
-        if (name != null && name.equalsIgnoreCase("broker")) {
-            logger.debug("Using the broker report downloader");
-            reportsDownloader = new BrokerDownloader(maestro, baseDir);
-        } else {
-            reportsDownloader = new DefaultDownloader(baseDir);
-            logger.debug("Using the default (HTTP) report downloader");
-        }
+        logger.debug("Using the broker report downloader");
+        reportsDownloader = new BrokerDownloader(maestro, baseDir);
 
         return pool ? new PooledDownloaderDecorator(reportsDownloader) : reportsDownloader;
     }
@@ -59,11 +54,7 @@ public class DownloaderBuilder {
         }
 
         final ReportsDownloader reportsDownloader;
-        if (name != null && name.equalsIgnoreCase("broker")) {
-            reportsDownloader = new BrokerDownloader(maestro, organizer);
-        } else {
-            reportsDownloader = new DefaultDownloader(organizer);
-        }
+        reportsDownloader = new BrokerDownloader(maestro, organizer);
 
         return pool ? new PooledDownloaderDecorator(reportsDownloader) : reportsDownloader;
     }
