+++ maestro-test-scripts/src/main/groovy/multipoint/FixedRateTestWithAgent.groovy
@@ -122,7 +122,6 @@ testProfile.setExtPointBranch(extPointBranch)
 testProfile.setExtPointCommand(extPointCommand)
 
 ManagementInterface.setupInterface(managementInterface, inspectorName, testProfile)
-ManagementInterface.setupResolver(inspectorName, reportsDownloader)
 
 FixedRateTestExecutor testExecutor = new FixedRateTestExecutor(maestro, reportsDownloader, testProfile,
         distributionStrategy)
