+++ maestro-test-scripts/src/main/groovy/singlepoint/FairIncrementalTest.groovy
@@ -171,7 +171,6 @@ if (maxLatencyStr != null) {
 }
 
 ManagementInterface.setupInterface(managementInterface, inspectorName, testProfile)
-ManagementInterface.setupResolver(inspectorName, reportsDownloader)
 
 IncrementalTestExecutor testExecutor = new IncrementalTestExecutor(maestro, reportsDownloader, testProfile, distributionStrategy)
