+++ maestro-test-scripts/src/main/groovy/singlepoint/FixedRateTest.groovy
@@ -108,7 +108,6 @@ testProfile.setRate(Integer.parseInt(rate))
 testProfile.setParallelCount(Integer.parseInt(parallelCount))
 
 ManagementInterface.setupInterface(managementInterface, inspectorName, testProfile)
-ManagementInterface.setupResolver(inspectorName, reportsDownloader)
 
 AbstractFixedRateExecutor testExecutor = FixedRateTestExecutorFactory.newTestExecutor(maestro, reportsDownloader,
         testProfile, warmUp, distributionStrategy)
