+++ maestro-test-scripts/src/main/groovy/singlepoint/IncrementalTest.groovy
@@ -140,7 +140,6 @@ testProfile.setInitialParallelCount(Integer.parseInt(initialParallelCount))
 testProfile.setCeilingParallelCount(Integer.parseInt(ceilingParallelCount))
 
 ManagementInterface.setupInterface(managementInterface, inspectorName, testProfile)
-ManagementInterface.setupResolver(inspectorName, reportsDownloader)
 
 IncrementalTestExecutor testExecutor = new IncrementalTestExecutor(maestro, reportsDownloader, testProfile, distributionStrategy)
